requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
psycopg2==2.9.5
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7
sqlparse==0.2.4
tzdata==2022.7
Werkzeug==2.2.2
